#2: [create] src directory
