+++ cmake-build-debug/CMakeFiles/clion-Debug-log.txt
@@ -1,0 +1,1 @@
+CMakeLists.txt not found in /Users/haru/42tokyo/42-philosophers Select CMakeLists.txt
